+++ src/Capsule.php
@@ -4,46 +4,15 @@ declare(strict_types=1);
 
 namespace PhpCfdi\XmlCancelacion;
 
-use Countable;
 use DateTimeImmutable;
+use PhpCfdi\XmlCancelacion\Cancellation\CancellationCapsule;
 
-class Capsule implements Countable
+/** @deprecated */
+class Capsule extends CancellationCapsule
 {
-    /** @var string */
-    private $rfc;
-
-    /** @var DateTimeImmutable */
-    private $date;
-
-    /** @var array This is a B-Tree array, values are stored in keys */
-    private $uuids;
-
     public function __construct(string $rfc, array $uuids = [], DateTimeImmutable $date = null)
     {
-        $this->rfc = $rfc;
-        $this->date = $date ?? new DateTimeImmutable('now');
-        foreach ($uuids as $uuid) {
-            $this->uuids[strtoupper($uuid)] = true;
-        }
-    }
-
-    public function rfc(): string
-    {
-        return $this->rfc;
-    }
-
-    public function date(): DateTimeImmutable
-    {
-        return $this->date;
-    }
-
-    public function uuids(): array
-    {
-        return array_keys($this->uuids);
-    }
-
-    public function count(): int
-    {
-        return count($this->uuids);
+        parent::__construct($rfc, $uuids, $date);
+        trigger_error(sprintf('Deprecated class since 0.5.0, use %s', CancellationCapsule::class), E_USER_DEPRECATED);
     }
 }