+++ src/Contracts/CapsuleInterface.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PhpCfdi\XmlCancelacion\Contracts;
+
+use DOMDocument;
+
+interface CapsuleInterface
+{
+    public function exportToDocument(): DOMDocument;
+
+    public function belongsToRfc(string $rfc): bool;
+}