+++ src/Exceptions/XmlCancelacionLogicException.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PhpCfdi\XmlCancelacion\Exceptions;
+
+use LogicException;
+
+class XmlCancelacionLogicException extends LogicException implements XmlCancelacionException
+{
+}