+++ tests/System/SignerImplementations/ImplementationDOMSignerTest.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PhpCfdi\XmlCancelacion\Tests\System\SignerImplementations;
+
+use PhpCfdi\XmlCancelacion\Contracts\SignerInterface;
+use PhpCfdi\XmlCancelacion\DOMSigner;
+
+class ImplementationDOMSignerTest extends SignerImplementationTestCase
+{
+    public function createSigner(): SignerInterface
+    {
+        return new DOMSigner();
+    }
+}